--- .gitlab-ci.yml
@@ -1,105 +1,0 @@
-# Reihenfolge für die Ausführung der Schritte
-# Bitte keine Stages entfernen, da sonst die Syntax u.U. ungültig wird!
-stages:
-  - codecheck
-  - build
-  - test
-  - dependency-check
-  - sonar
-  - deploy
-  - release
-  - build-image
-  - review
-  - cleanup
-  - deploy-dev
-  - integration-test
-  - deploy-test
-  - system-test
-  - deploy-production
-
-# Auswahl der Runner
-default:
-  tags:
-    - openshift
-
-variables:
-  # Hinweis: Variable APP_JAR_FILE wird in build.mvn ermittelt, als Umgebungsvariable hinzugefügt und in den mvn-Jobs genutzt
-  # Basispfad der Anwendung. Wird für den Dateipfad zur JAR-Datei benötigt.
-  APP_PATH: .
-  # wenn ein Artifact Deployment stattfinden soll, dann ausklammern.
-  DISABLE_DEPLOYMENT_ARTIFACTORY: "true"
-  # Leerzeichen-separierte Liste von SAST Analyzers, welche nicht ausgeführt werden sollen
-  # Siehe https://gitlab.com/gitlab-org/gitlab/-/blob/master/lib/gitlab/ci/templates/Jobs/SAST.gitlab-ci.yml
-  SAST_EXCLUDED_ANALYZERS: "eslint nodejs-scan semgrep spotbugs"
-  # Name für Openshift Objekte DeploymentConfig, BuildConfig, ConfigMap, ImageStream
-  OPENSHIFT_APP_NAME:
-  # Openshift Namespace der DEV-Umgebung.
-  OPENSHIFT_DEV_NAMESPACE:
-  # Umgebungsname für die DEV-Umgebung (wird an DC, BC, CM an den Namen angehängt).
-  OPENSHIFT_DEV_ENV: dev
-  # Openshift Namespace der TEST-Umgebung.
-  OPENSHIFT_TEST_NAMESPACE:
-  # Umgebungsname für die TEST-Umgebung (wird an DC, BC, CM an den Namen angehängt).
-  OPENSHIFT_TEST_ENV: test
-  # Openshift Namespace der PROD-Umgebung.
-  OPENSHIFT_PROD_NAMESPACE:
-  # Umgebungsname für die PROD-Umgebung (wird an DC, BC, CM an den Namen angehängt).
-  OPENSHIFT_PROD_ENV: prod
-
-# Legt das Verhalten in Pipeline in verschiedenen Fällen fest:
-# 1. Der Auslöser ist ein Bot und es ist keine Tag-Pipeline:
-#    Die Pipeline wird nicht gestartet, da es sich hierbei um eine Versionsänderung duch das Maven Release Plugin handelt
-# 2. DO_RELEASE ist nicht auf "true" und der Commit kommt von einem Branch oder Tag:
-#    Das ist der "Normalfall" und die Pipeline wird mit allen notwendigen Variablen gestartet
-# 3. DO_RELEASE ist auf "true" und der Commit kommt von einem Branch oder Tag:
-#    Die Pipeline wird ohne Variablen gestartet, da sie lediglich einen Release durchführen soll
-workflow:
-  rules:
-    - if: $CI_PIPELINE_SOURCE  == "schedule"
-      variables:
-        DISABLE_SELENIUM: "true"
-        DISABLE_OPENSHIFT: "true"
-        DISABLE_DEPLOYMENT_ARTIFACTORY: "true"
-    - if: $GITLAB_USER_LOGIN =~ /^project_/ && $CI_COMMIT_TAG == null
-      when: never
-    - if: $DO_RELEASE != "true" && ($CI_COMMIT_TAG || $CI_COMMIT_BRANCH)
-
-    - if: $DO_RELEASE == "true" && ($CI_COMMIT_TAG || $CI_COMMIT_BRANCH)
-      variables:
-        DISABLE_OPENSHIFT: "true"
-        DISABLE_SONAR: "true"
-        DISABLE_OWASP: "true"
-        DISABLE_DEPLOYMENT_ARTIFACTORY: "true"
-
-include:
-  # Für Branch Pipelines
-  - template: "Workflows/Branch-Pipelines.gitlab-ci.yml"
-
-  - project: "ccse/cicd/gitlab-runner-templates"
-    ref: "main-2.x"
-    file:
-      # Gitlab AutoDevOps Secret Detection (siehe: https://docs.gitlab.com/ee/topics/autodevops/stages.html#auto-secret-detection)
-      - "SecretDetection.gitlab-ci.yml"
-      # Gitlab AutoDevOps SAST (siehe: https://docs.gitlab.com/ee/topics/autodevops/stages.html#auto-sast)
-      - "SAST.gitlab-ci.yml"
-      # Maven Build (stage: build), Unit-Tests (stage: test) und Artifactory Deployment (stage: deploy-dev)
-      - "Maven.gitlab-ci.yml"
-      # Maven Release Plugin
-      - "MavenRelease.gitlab-ci.yml"
-      # Codecheck Analyse
-      - "Codecheck.gitlab-ci.yml"
-      # Include Sonarqube Step
-      - "SonarQube.gitlab-ci.yml"
-      # OWASP Dependency Check
-      - OWASP-Dependency-Check.gitlab-ci.yml
-      # Openshift Image Build über S2I und Deployment in Dev (stage: deploy-dev)
-      - "Openshift-CI.gitlab-ci.yml"
-      # Deployment (Rollout) in Test (stage: deploy-test) und in Prod (stage: deploy-prod)
-      - "Openshift-CD.gitlab-ci.yml"
-
-# Der Dependency-Check analysiert verwendete Libraries auf bekannte Schwachstellen.
-# SUPPRESSION_FILE kann erweitert werden um eine eigene Custom-Suppressions-Datei,
-# indem diese mit Leerzeichen getrennt dahinter geschrieben werden.
-dependency-check:
-  variables:
-    SUPPRESSION_FILE: "--suppression https://ccse.pages.muenchen.de/refarch-security/backend/suppression-cves.xml"
