+++ src/main/java/de/muenchen/mobidam/mobilithek/InterfaceMessageFactory.java
@@ -25,6 +25,7 @@ package de.muenchen.mobidam.mobilithek;
 import de.muenchen.mobidam.Constants;
 import de.muenchen.mobidam.config.Interfaces;
 import de.muenchen.mobidam.integration.client.domain.DatentransferCreateDTO;
+import de.muenchen.mobidam.sstmanagment.EreignisTyp;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import org.apache.camel.Exchange;
@@ -43,7 +44,7 @@ public class InterfaceMessageFactory {
         dto.setEreignis(EreignisTyp.BEGINN.name());
         dto.setZeitstempel(LocalDateTime.now());
         dto.setProzessId(exchange.getExchangeId());
-        dto.setSchnittstelle(interfaceProps.getInterfaces().get(Constants.PARK_RIDE_DATA).getMobidamSstId());
+        dto.setSchnittstelle(interfaceProps.getInterfaces().get(Constants.PARK_RIDE_STATIC_DATA).getMobidamSstId());
         exchange.getMessage().setBody(dto);
     }
 
@@ -53,7 +54,7 @@ public class InterfaceMessageFactory {
         dto.setEreignis(EreignisTyp.ERFOLG.name());
         dto.setZeitstempel(LocalDateTime.now());
         dto.setProzessId(exchange.getExchangeId());
-        dto.setSchnittstelle(interfaceProps.getInterfaces().get(Constants.PARK_RIDE_DATA).getMobidamSstId());
+        dto.setSchnittstelle(interfaceProps.getInterfaces().get(Constants.PARK_RIDE_STATIC_DATA).getMobidamSstId());
         dto.setInfo(String.format("Interface status code : '%s' (%s)", exchange.getIn().getHeader(Exchange.HTTP_RESPONSE_CODE),
                 exchange.getIn().getHeader(AWS2S3Constants.KEY)));
         exchange.getMessage().setBody(dto);
@@ -65,7 +66,7 @@ public class InterfaceMessageFactory {
         dto.setEreignis(EreignisTyp.FEHLER.name());
         dto.setZeitstempel(LocalDateTime.now());
         dto.setProzessId(exchange.getExchangeId());
-        dto.setSchnittstelle(interfaceProps.getInterfaces().get(Constants.PARK_RIDE_DATA).getMobidamSstId());
+        dto.setSchnittstelle(interfaceProps.getInterfaces().get(Constants.PARK_RIDE_STATIC_DATA).getMobidamSstId());
         var ex = exchange.getException() != null ? exchange.getException() : (Exception) exchange.getAllProperties().get(Exchange.EXCEPTION_CAUGHT);
         dto.setInfo(String.format("End interface with error : %s", ex.getMessage()));
         exchange.getMessage().setBody(dto);
@@ -77,7 +78,7 @@ public class InterfaceMessageFactory {
         dto.setEreignis(EreignisTyp.ENDE.name());
         dto.setZeitstempel(LocalDateTime.now());
         dto.setProzessId(exchange.getExchangeId());
-        dto.setSchnittstelle(interfaceProps.getInterfaces().get(Constants.PARK_RIDE_DATA).getMobidamSstId());
+        dto.setSchnittstelle(interfaceProps.getInterfaces().get(Constants.PARK_RIDE_STATIC_DATA).getMobidamSstId());
         exchange.getMessage().setBody(dto);
     }
