+++ src/main/java/de/muenchen/mobidam/mobilithek/MobilithekEaiRouteBuilder.java
@@ -62,11 +62,13 @@ public class MobilithekEaiRouteBuilder extends RouteBuilder {
                     .setHeader(Constants.PARAMETER_BUCKET_NAME, simple(String.format("${header.%s.s3Bucket}", Constants.INTERFACE_TYPE)))
                     .process("s3CredentialProvider")
                     .process("s3ObjectKeyBuilder")
-                    .toD("aws2-s3://$simple{header.bucketName}?accessKey=RAW($simple{header.accessKey})&secretKey=RAW($simple{header.secretKey})&region=${camel.component.aws2-s3.region}&overrideEndpoint=true&uriEndpointOverride=${camel.component.aws2-s3.override-endpoint}").id(MOBIDAM_ENDPOINT_S3_ID)
+                    .toD("aws2-s3://${header.bucketName}?accessKey=RAW(${header.accessKey})&secretKey=RAW(${header.secretKey})&region=${camel.component.aws2-s3.region}&overrideEndpoint=true&uriEndpointOverride=${camel.component.aws2-s3.override-endpoint}").id(MOBIDAM_ENDPOINT_S3_ID)
                     .bean("interfaceMessageFactory", "parkRideDataSuccess")
                     .bean("sstManagementIntegrationService", "logDatentransfer")
                     .bean("interfaceMessageFactory", "parkRideDataEnd")
                     .bean("sstManagementIntegrationService", "logDatentransfer")
+                .otherwise()
+                    .log(LoggingLevel.DEBUG, Constants.MOBIDAM_LOGGER, String.format("${header.%s.mobidamSstId} is not active.", Constants.INTERFACE_TYPE))
                 .end();
         //  spotless:on
