+++ src/main/resources/logback-spring.xml
@@ -1,0 +1,124 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+    The MIT License
+    Copyright © 2024 Landeshauptstadt München | it@M
+
+    Permission is hereby granted, free of charge, to any person obtaining a copy
+    of this software and associated documentation files (the "Software"), to deal
+    in the Software without restriction, including without limitation the rights
+    to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+    copies of the Software, and to permit persons to whom the Software is
+    furnished to do so, subject to the following conditions:
+
+    The above copyright notice and this permission notice shall be included in
+    all copies or substantial portions of the Software.
+
+    THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+    IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+    FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+    AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+    LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+    OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+    THE SOFTWARE.
+
+-->
+
+<!-- The possibly existing properties concerning the "logging.level" in the application.yml
+     have priority over the log level maintained in this file.  -->
+<configuration>
+
+
+    <springProperty scope="context" name="springAppName" source="spring.application.name"/>
+
+
+    <springProfile name="local,test">
+
+        <!-- Log appender -->
+        <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+            <encoder>
+                <pattern>
+                    %date{yyyy.MM.dd HH:mm:ss.SSS} | ${springAppName} | TraceId: %X{traceId:-} | SpanId: %X{spanId:-}] | %level | [%thread] | %logger{0} | [%file : %line] - %msg%n
+                </pattern>
+            </encoder>
+        </appender>
+
+        <!-- Logger -->
+        <logger name="de.muenchen.mobidam"
+                level="debug"
+                additivity="false">
+            <appender-ref ref="STDOUT" />
+        </logger>
+
+        <logger name="de.muenchen.allgemein"
+                level="debug"
+                additivity="false">
+            <appender-ref ref="STDOUT" />
+        </logger>
+
+        <logger name="org.springframework.cache"
+                level="trace"
+                additivity="false">
+            <appender-ref ref="STDOUT" />
+        </logger>
+
+        <!-- Logger usage -->
+        <root level="info">
+            <appender-ref ref="STDOUT" />
+        </root>
+
+    </springProfile>
+
+
+    <springProfile name="dev,kon,prod,openshift">
+
+        <!-- Log appender -->
+        <appender name="JSON_STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+            <encoder class="net.logstash.logback.encoder.LoggingEventCompositeJsonEncoder">
+                <providers>
+                    <!-- provides the fields in the configured pattern -->
+                    <pattern>
+                        <!-- the pattern that defines what to include -->
+                        <pattern>
+                            {
+                                "timestamp" : "%date{yyyy-MM-dd'T'HH:mm:ss.SSS}",
+                                "appName" : "${springAppName}",
+                                "TraceId" : "%mdc{traceId}",
+                                "SpanId" : "%mdc{spanId}",
+                                "X-Span-Export" : "%mdc{X-Span-Export}",
+                                "thread" : "%thread",
+                                "level" : "%level",
+                                "logger": "%logger",
+                                "location" : {
+                                    "fileName" : "%file",
+                                    "line" : "%line"
+                                },
+                                "message": "%message"
+                            }
+                        </pattern>
+                    </pattern>
+                </providers>
+            </encoder>
+        </appender>
+
+        <!-- Logger -->
+        <logger name="de.muenchen.mobidam"
+                level="info"
+                additivity="false">
+            <appender-ref ref="JSON_STDOUT" />
+        </logger>
+
+        <logger name="de.muenchen.allgemein"
+                level="info"
+                additivity="false">
+            <appender-ref ref="JSON_STDOUT" />
+        </logger>
+
+        <!-- Logger usage -->
+        <root level="info">
+            <appender-ref ref="JSON_STDOUT" />
+        </root>
+
+    </springProfile>
+
+</configuration>
