--- src/main/resources/logback.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    The MIT License
-    Copyright © 2024 Landeshauptstadt München | it@M
-
-    Permission is hereby granted, free of charge, to any person obtaining a copy
-    of this software and associated documentation files (the "Software"), to deal
-    in the Software without restriction, including without limitation the rights
-    to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-    copies of the Software, and to permit persons to whom the Software is
-    furnished to do so, subject to the following conditions:
-
-    The above copyright notice and this permission notice shall be included in
-    all copies or substantial portions of the Software.
-
-    THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-    IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-    FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-    AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-    LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-    OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-    THE SOFTWARE.
-
--->
-<configuration>
-
-    <appender name="ConsoleAppender" class="ch.qos.logback.core.ConsoleAppender">
-        <layout class="ch.qos.logback.classic.PatternLayout">
-            <Pattern>
-                %d [%t] %-5level %logger{36} - %msg%n   <!-- http://logback.qos.ch/manual/layouts.html -->
-            </Pattern>
-        </layout>
-    </appender>
-
-    <root level="info">
-        <appender-ref ref="ConsoleAppender" />
-    </root>
-    <logger name="de.muenchen" level="all" additivity="false">
-        <appender-ref ref="ConsoleAppender" />
-    </logger>
-
-</configuration>
