+++ src/test/resources/application.yml
@@ -23,8 +23,8 @@ camel:
 de.muenchen.mobidam.integration:
   baseUrl: https://...
   interfaces:
-    parkRideData:
-      mobidam-sst-id: 904fcf2d-25bb-4fa9-85ff-f7ed12348fe4
+    parkRideStaticData:
+      mobidam-sst-id: 999fcf2d-25bb-4fa9-85ff-f7ed12349999
       mobilithek-subscription-id: 123456789
       mobilithek-url: mock:mobilithekInfo
       cron-expression: '0 0 0 ? * *'
@@ -34,7 +34,6 @@ de.muenchen.mobidam.integration:
 
 mobidam:
   s3:
-    url: mock:s3Destination
     bucket-credential-config:
       int-mdasc-mdasdev:
         access-key-env-var: MOBIDAM_INT-MDASC-MDASDEV_ACCESS_KEY
